[WIP] shuffle around, more parsing.
